move app to componets
