--- src/App.css
@@ -1,3 +1,0 @@
-li {
-    list-style: none;
-}
