--- src/app.module.css
@@ -1,6 +1,0 @@
-.main {    
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
