+++ src/components/app/app.module.css
@@ -1,0 +1,6 @@
+.main {    
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
