+++ src/index.css
@@ -1,0 +1,3 @@
+li {
+    list-style: none;
+}
